feat: add `dictionary` feature
Introduce the new `dictionary` feature. This feature provides the `Dictionary` class and the `dictionary` module.

Adds the `keys`, `values`, `entries`, `size`, `at`, and `set` utilities to the `Dictionary` class, also adds to the `dictionary` module its own versions of these utilities.

Implements the respective unit tests to the new utilities in both elements.